cleanup: remove unused db engines
