+++ crates/storage/Cargo.toml
@@ -24,8 +24,6 @@ rocksdb = { version = "0.19.0", optional = true }
 default = ["in_memory", "libmdbx"]
 in_memory = []
 libmdbx = ["dep:libmdbx"]
-sled = ["dep:sled"]
-rocksdb = ["dep:rocksdb"]
 
 [dev-dependencies]
 hex.workspace = true
