+++ crates/storage/engines/api.rs
@@ -1,0 +1,152 @@
+use std::fmt::Debug;
+
+use bytes::Bytes;
+use ethereum_types::{Address, H256};
+
+use ethereum_rust_core::types::{
+    Account, AccountInfo, BlockBody, BlockHash, BlockHeader, BlockNumber, Index, Receipt,
+    Transaction,
+};
+
+use crate::error::StoreError;
+
+pub(crate) type Key = Vec<u8>;
+pub(crate) type Value = Vec<u8>;
+
+pub trait StoreEngine: Debug + Send {
+    /// Add account info
+    fn add_account_info(
+        &mut self,
+        address: Address,
+        account_info: AccountInfo,
+    ) -> Result<(), StoreError>;
+
+    /// Obtain account info
+    fn get_account_info(&self, address: Address) -> Result<Option<AccountInfo>, StoreError>;
+
+    /// Add block header
+    fn add_block_header(
+        &mut self,
+        block_number: BlockNumber,
+        block_header: BlockHeader,
+    ) -> Result<(), StoreError>;
+
+    /// Obtain block header
+    fn get_block_header(
+        &self,
+        block_number: BlockNumber,
+    ) -> Result<Option<BlockHeader>, StoreError>;
+
+    /// Add block body
+    fn add_block_body(
+        &mut self,
+        block_number: BlockNumber,
+        block_body: BlockBody,
+    ) -> Result<(), StoreError>;
+
+    /// Obtain block body
+    fn get_block_body(&self, block_number: BlockNumber) -> Result<Option<BlockBody>, StoreError>;
+
+    /// Add block body
+    fn add_block_number(
+        &mut self,
+        block_hash: BlockHash,
+        block_number: BlockNumber,
+    ) -> Result<(), StoreError>;
+
+    /// Obtain block number
+    fn get_block_number(&self, block_hash: BlockHash) -> Result<Option<BlockNumber>, StoreError>;
+
+    /// Store transaction location (block number and index of the transaction within the block)
+    fn add_transaction_location(
+        &mut self,
+        transaction_hash: H256,
+        block_number: BlockNumber,
+        index: Index,
+    ) -> Result<(), StoreError>;
+
+    /// Obtain transaction location (block number and index)
+    fn get_transaction_location(
+        &self,
+        transaction_hash: H256,
+    ) -> Result<Option<(BlockNumber, Index)>, StoreError>;
+
+    /// Add receipt
+    fn add_receipt(
+        &mut self,
+        block_number: BlockNumber,
+        index: Index,
+        receipt: Receipt,
+    ) -> Result<(), StoreError>;
+
+    /// Obtain receipt
+    fn get_receipt(
+        &self,
+        block_number: BlockNumber,
+        index: Index,
+    ) -> Result<Option<Receipt>, StoreError>;
+
+    /// Set an arbitrary value (used for eventual persistent values: eg. current_block_height)
+    fn set_value(&mut self, key: Key, value: Value) -> Result<(), StoreError>;
+
+    /// Retrieve a stored value under Key
+    fn get_value(&self, key: Key) -> Result<Option<Value>, StoreError>;
+
+    /// Add account code
+    fn add_account_code(&mut self, code_hash: H256, code: Bytes) -> Result<(), StoreError>;
+
+    /// Obtain account code via code hash
+    fn get_account_code(&self, code_hash: H256) -> Result<Option<Bytes>, StoreError>;
+
+    /// Obtain account code via account address
+    fn get_code_by_account_address(&self, address: Address) -> Result<Option<Bytes>, StoreError> {
+        let code_hash = match self.get_account_info(address)? {
+            Some(acc_info) => acc_info.code_hash,
+            None => return Ok(None),
+        };
+        self.get_account_code(code_hash)
+    }
+
+    fn get_transaction_by_hash(
+        &self,
+        transaction_hash: H256,
+    ) -> Result<Option<Transaction>, StoreError> {
+        let (block_number, index) = match self.get_transaction_location(transaction_hash)? {
+            Some(locations) => locations,
+            None => return Ok(None),
+        };
+        let block_body = match self.get_block_body(block_number)? {
+            Some(body) => body,
+            None => return Ok(None),
+        };
+        Ok(index
+            .try_into()
+            .ok()
+            .and_then(|index: usize| block_body.transactions.get(index).cloned()))
+    }
+
+    // Add storage value
+    fn add_storage_at(
+        &mut self,
+        address: Address,
+        storage_key: H256,
+        storage_value: H256,
+    ) -> Result<(), StoreError>;
+
+    // Obtain storage value
+    fn get_storage_at(
+        &self,
+        address: Address,
+        storage_key: H256,
+    ) -> Result<Option<H256>, StoreError>;
+
+    /// Stores account in db (including info, code & storage)
+    fn add_account(&mut self, address: Address, account: Account) -> Result<(), StoreError> {
+        self.add_account_info(address, account.info.clone())?;
+        self.add_account_code(account.info.code_hash, account.code)?;
+        for (storage_key, storage_value) in account.storage {
+            self.add_storage_at(address, storage_key, storage_value)?;
+        }
+        Ok(())
+    }
+}
